Refactor Butterfly Vision: extract parallax logic to `useParallaxBackground`, enhance HeroSection structure, and update SeparatorSection styling.
